app/api/requirements.txt
fastapi==0.82.0
python-dotenv==0.21.0
pymongo[srv]==4.2.0
uvicorn==0.18.3